Fix warnings.

+++ src/lib_wnd/input.cpp
@@ -75,7 +75,7 @@ EXPORT_CPP void _setCfgKey(S64 idx, S64 btn, const U8* keys)
 
 EXPORT_CPP Bool _getKey(S64 key)
 {
-	return (GetKeyState(key) & 0x8000) != 0;
+	return (GetKeyState(static_cast<int>(key)) & 0x8000) != 0;
 }
 
 namespace Input
